requirements.txt
streamlit
tensorflow==2.12.1
numpy
Pillow